Create  Notebook: Adding dates to the system
In this exercise, add a feature which includes the date and time of the written note to the program. The program works as earlier, but saves data in the form "[note]:::[date and time]" meaning that there is a three-colon separator between the note and timestamp. The timestamp can be generated as follows:




	
>>> import time		

>>> time.strftime("%X %x")
'19:01:34 01/03/09'
>>> 

This returns the date and time in a nice, compact string. When working correctly, the program prints the following:




			
>>> 
(1) Read the notebook
(2) Add note
(3) Empty the notebook
(4) Quit

Please select one: 2
Write a new note: Call mom.
(1) Read the notebook
(2) Add note
(3) Empty the notebook
(4) Quit

Please select one: 1
Call mom.:::11:44:41 04/25/11

(1) Read the notebook
(2) Add note
(3) Empty the notebook
(4) Quit

Please select one: 4
Notebook shutting down, thank you.
>>>